+++ src/python/zquantum/core/circuits/__init__.py
@@ -141,9 +141,7 @@ from ._builtin_gates import (
     builtin_gate_by_name,
 )
 from ._circuit import Circuit
-from ._cirq_decompositions import PowerGateToPhaseAndRotation, decompose_cirq_circuit
 from ._compatibility import new_circuit_from_old_circuit
-from ._decomposition import DecompositionRule, decompose_operations
 from ._gates import (
     ControlledGate,
     CustomGateDefinition,
