+++ src/python/zquantum/core/decompositions/__init__.py
@@ -1,0 +1,2 @@
+from ._decomposition import decompose_operations
+from ._cirq_decompositions import decompose_cirq_circuit, PowerGateToPhaseAndRotation
