+++ src/python/zquantum/core/decompositions/_decomposition.py
